Fix Utils.cloneImageData throwing an error in older browsers
In some old browsers, image data was in a `CanvasPixelArray` (instead of a `Uint8ClampedArray`) that did not have a `set` function.  This adds a fallback of manually copying the data.  At least old Safari does not define `CanvasPixelArray` globally, so it is not as straightforward to simply polyfill `CanvasPixelArray.prototype.set`.