Refactor curve tool to update when colors and line width are changed
This changes the curve tool state to reflect the point that started being placed on last pointer down instead of the point placed on last pointer up.  Since nothing increases the pointer state other than pointer down, that *should* not be an issue.  That leaves little to be done on pointer up.  This refactor also combines how some tool states are handled since they are functionally the same.  It also fixes #253.